Move wallet content to dashboard; removed wallet page. Renamed logout/login to sign in/sign out. Added public key auth context var in settings page.
